Add CLI flags for easier unit testing, minor fixes, update README

+++ src/main.rs
@@ -1,15 +1,21 @@
 use std::env;
 use std::fs;
-use std::process::Command;
 use std::io::Write;
+use std::process::Command;
 
-mod lexer; use lexer::Lexer;
-mod parser; use parser::Parser;
-mod codegen; use codegen::AssemblyGenerator;
+mod lexer;
+use lexer::Lexer;
+mod parser;
+use parser::Parser;
+mod codegen;
+use codegen::AssemblyGenerator;
 
 struct Config {
     infiles: Vec<String>,
     outfile: Option<String>,
+    run_parser: bool,
+    run_codegen: bool,
+    emit_code: bool,
 }
 
 // TODO: better file handling
@@ -19,10 +25,13 @@ impl Config {
         Config {
             infiles: Vec::new(),
             outfile: None,
+            run_parser: true,
+            run_codegen: true,
+            emit_code: true,
         }
     }
 
-    fn parse() -> Result<Config, &'static str> {
+    fn parse() -> Result<Config, String> {
         let mut cfg = Config::new();
         let mut args = env::args().collect::<Vec<String>>().into_iter().skip(1);
 
@@ -33,10 +42,22 @@ impl Config {
                         if let Some(file) = args.next() {
                             cfg.outfile = Some(file)
                         } else {
-                            return Err("expected file following output flag");
+                            return Err(format!("expected file following `{}`", arg));
                         }
                     }
-                    _ => return Err("illegal option"),
+                    "--lex" => {
+                        cfg.run_parser = false;
+                        cfg.run_codegen = false;
+                        cfg.emit_code = false;
+                    }
+                    "--parse" => {
+                        cfg.run_codegen = false;
+                        cfg.emit_code = false;
+                    }
+                    "--codegen" => {
+                        cfg.emit_code = false;
+                    }
+                    other => return Err(format!("illegal flag `{other}`")),
                 }
             } else {
                 cfg.infiles.push(arg)
@@ -49,31 +70,31 @@ impl Config {
 fn preprocess(infiles: &Vec<String>, outfile: &String) {
     // TODO: multiple outfiles
     Command::new("gcc")
-        .args([
-            "-E",
-            "-P",
-            &infiles.join(" "),
-            "-o",
-            outfile,
-        ])
+        .args(["-E", "-P", &infiles.join(" "), "-o", outfile])
         .output()
         .expect("preprocessing failed");
 }
 
-fn compile(infile: &String, outfile: &String) {
-    let code = fs::read_to_string(infile)
-        .expect("Failed to read file");
+fn compile(cfg: &Config, infile: &String, outfile: &String) -> Result<(), Box<dyn std::error::Error>>{
+    let code = fs::read_to_string(infile).expect("Failed to read file");
 
     let tokens = Lexer::new(code.as_bytes()).get_tokens();
 
-    let program = Parser::new(tokens).parse_program().unwrap();
+    if cfg.run_parser {
+        let program = Parser::new(tokens).parse_program()?;
 
-    let codegen = AssemblyGenerator::new();
-    let translated = codegen.translate(program);
-    let asm = codegen.generate_asm(translated);
+        if cfg.run_codegen {
+            let codegen = AssemblyGenerator::new();
+            let translated = codegen.translate(program);
+            let asm = codegen.generate_asm(translated);
 
-    let mut file = fs::File::create(outfile).expect("Failed to create output file");
-    write!(file, "{}", asm).expect("Failed to write to output file");
+            if cfg.emit_code {
+                let mut file = fs::File::create(outfile).expect("Failed to create output file");
+                write!(file, "{}", asm).expect("Failed to write to output file");
+            }
+        }
+    }
+    Ok(())
 }
 
 fn assemble_and_link(assembly_file: &String, outfile: &String) {
@@ -83,24 +104,30 @@ fn assemble_and_link(assembly_file: &String, outfile: &String) {
         .expect("assembling/linking failed");
 }
 
-fn run(cfg: Config) {
+fn run(cfg: Config) -> Result<(), Box<dyn std::error::Error>> {
     // TODO: better file handling (PathBuf?)
-    // let binding = cfg.outfile.clone().expect("no output file");
-    // let output = binding;
-    let output = cfg.outfile.clone().expect("no output file");
+    let output = match cfg.outfile {
+        Some(ref s) => s.clone(),
+        None => "a.out".to_string(),
+    };
+    // let output = cfg.outfile.clone().expect("no output file");
     let preprocessor_output = output.to_owned() + ".i";
     let assembly_output = output.to_owned() + ".s";
 
     preprocess(&cfg.infiles, &preprocessor_output);
 
-    compile(&preprocessor_output, &assembly_output);
+    compile(&cfg, &preprocessor_output, &assembly_output)?;
     fs::remove_file(preprocessor_output).expect("failed to remove preprocessed file");
 
-    assemble_and_link(&assembly_output, &output);
-    fs::remove_file(assembly_output).expect("failed to remove assembly file");
+    if cfg.emit_code {
+        assemble_and_link(&assembly_output, &output);
+        fs::remove_file(assembly_output).expect("failed to remove assembly file");
+    }
+    Ok(())
 }
 
-fn main() {
-    let cfg = Config::parse().unwrap();
-    run(cfg);
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let cfg = Config::parse()?;
+    run(cfg)?;
+    Ok(())
 }
